feat: encode referralProjectId as (chainId, projectId) pair; null on 0
- `referralProjectId` is now a packed `(chainId << 48) | projectId` reference,
  so a referrer with a project on chain X can be credited for fee-paying
  activity on chain Y. Bare project IDs (chain bits zero) still work as
  "current chain" — the natural default.
- Repack `JBFee` to keep the struct at two storage slots: shrink `amount` to
  `uint224` (consistent with the protocol's payout/surplus-allowance limits)
  and add `uint32 referralChainId` beside it. Held-fee processing reconstructs
  the encoded transient `(chainId << 48) | projectId` from the two halves.
- Drop the "pass 0 → credit the project itself" defaulting in `cashOutTokensOf`,
  `sendPayoutsOf`, and `useAllowanceOf`. `0` now means "no referral credit",
  matching the null semantics expected by indexers.
- Document the encoding in `JBFee`, the transient `currentReferralProjectId`,
  `IJBMultiTerminal.currentReferralProjectId`, `IJBTerminalStore.feeVolumeByReferralOf`,
  and the three entry-point param docs.
- Update `TestProcessHeldFeesOf` storage planter + struct literals for the new
  two-half layout.

+++ src/libraries/JBHeldFees.sol
@@ -88,7 +88,11 @@ library JBHeldFees {
                 // refunded net and its fee, leaving the slot live so the leftover can be refunded later.
                 feeAmount = JBFees.standardFeeAmountResultingIn(leftoverAmount);
                 unchecked {
-                    heldFeesOf[projectId][token][i].amount -= (leftoverAmount + feeAmount);
+                    // `JBFee.amount` is `uint224`; the subtraction operand is `uint256`. Narrow the operand to
+                    // `uint224` for the in-place update. The held gross amount was itself stored as `uint224`
+                    // and the partial refund is bounded above by it, so the narrowed subtrahend always fits.
+                    // forge-lint: disable-next-line(unsafe-typecast)
+                    heldFeesOf[projectId][token][i].amount -= uint224(leftoverAmount + feeAmount);
                     returnedFees += feeAmount;
                 }
                 // All of the incoming amount has been matched — exit the loop next iteration.
